working varaint extraction, missing tests
