+++ package.json
@@ -77,6 +77,7 @@
       "SceneNode",
       "PageNode",
       "RectangleNode",
+      "ComponentSetNode",
       "ComponentNode",
       "GridLayoutGrid",
       "RowsColsLayoutGrid",
